mostly converted dispatcher to TypeScript
