+++ test/server/koa-middleware.test.js
@@ -1,5 +1,5 @@
 import { Registry } from "../../src/server/registry.ts";
-import { Dispatcher } from "../../src/server/dispatcher.js";
+import { Dispatcher } from "../../src/server/dispatcher.ts";
 import { koaMiddleware } from "../../src/server/koa-middleware.js";
 import { ContextRegistry } from "../../src/server/context-registry.ts";
